+++ src/Ext/WarheadType/Hooks.cpp
@@ -112,7 +112,7 @@ void ApplyLogics(WarheadTypeClass* pWH , WeaponTypeClass*pWeapon ,BulletClass * 
 		if (pTypeExt->ReturnWeapon && pThis->Owner && pThis->Owner->IsAlive)
 		{
 			auto const RpWeapon = pTypeExt->ReturnWeapon.Get();
-			int damage = TechnoExtData::GetDamageMult(pThis->Owner  , RpWeapon->Damage, !pTypeExt->ReturnWeapon_ApplyFirepowerMult);
+			int damage = (int)TechnoExtData::GetDamageMult(pThis->Owner  , RpWeapon->Damage, !pTypeExt->ReturnWeapon_ApplyFirepowerMult);
 
 			if (BulletClass* pBullet = RpWeapon->Projectile->CreateBullet(pThis->Owner, pThis->Owner,
 				damage, RpWeapon->Warhead, RpWeapon->Speed, RpWeapon->Bright))
